requirements.txt
decorator==4.0.10
Django==1.10.1
django-tastypie==0.13.3
extras==1.0.0
fixtures==3.0.0
gevent==1.1.2
greenlet==0.4.10
linecache2==1.0.0
pbr==1.10.0
PyContracts==1.7.9
pyparsing==2.1.8
pysynthetic==0.4.14
python-dateutil==2.5.3
python-mimeparse==1.5.2
python-subunit==1.2.0
six==1.10.0
testrepository==0.0.20
testtools==2.2.0
traceback2==1.4.0
unittest2==1.1.0
